particles gravity, collision shapes, direction flipping, crouch fixed, slide transitions, sprite centered
tons of things, particles gravity set to 0, collision shapes values and shapes added for every animation including the new sliding collision shape, direction of the sprite fixed to the use of a bool instead of flipping the sprite, flip the whole character, position of the sprite centered for every single animation and it also works with the flipping, crouch walk fixed, sliding added transitions and new collision shape.

still needed logic on when to stand form crouching and sliding.